requirements.txt
gradio
requests
python-dotenv
faiss-cpu
numpy
PyPDF2
gTTS
